src/reports/requirements.txt
psycopg2-binary
ujson
